requirements.txt
streamlit==1.32.0
mindee==3.7.0
google-generativeai==0.3.2
pandas==2.2.0
matplotlib==3.8.2
plotly==5.18.0
python-dotenv==1.0.1 